requirements.txt
telethon
python-dotenv
langchain-groq
langchain
